move logging layer into prependable module
